requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==4.2.4
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
docx2pdf==0.1.8
docx2python==2.8.0
docx2txt==0.8
exceptiongroup==1.2.0
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
grpcio==1.62.1
grpcio-status==1.48.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.13.7
httpx==0.20.0
hyperframe==5.2.0
idna==2.10
lxml==5.1.0
openai==1.12.0
pillow==10.2.0
plum-dispatch==1.7.4
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.1
pywin32==306
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
Spire.Doc==12.2.1
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1